Add inline annotated function disassembly
These are Ghidra disassemblies of functions where
the DWARF data indicates functions where inlined.
Instructions belonging to an inlined function are
marked as such at the end. This is particularly
helpful for functions that access peripheral
registers where helper functions are used that
are almost always inlined and the name of the
inlined function can reveal the register name
or purpose.